Fixing a bug in the slider Fixing bugs in the joystick

+++ controls/player/PlayerViewModel.cs
@@ -28,7 +28,7 @@ namespace ViewModel
             get
             {
                 TimeSpan diff = model.Time - DateTime.MinValue;
-                return (float)diff.TotalSeconds;
+                return (int)diff.TotalSeconds;
             }
             set
             {
